TM-1314 merge contentOverviewProject into contentOverview
