+++ src/pages/dashboard/index.page.tsx
@@ -106,7 +106,7 @@ const Dashboard = () => {
     : [];
 
   return (
-    <div className="mt-4 mb-4 mr-2 flex flex-1 flex-wrap gap-4 overflow-auto bg-neutral-70 pl-4 pr-2 small:flex-nowrap">
+    <div className="mb-4 mr-2 mt-4 flex flex-1 flex-wrap gap-4 overflow-auto bg-neutral-70 pl-4 pr-2 small:flex-nowrap">
       <div className="overflow-hiden mx-auto w-full max-w-[730px] small:w-1/2 small:max-w-max">
         <PageRow className="gap-4 p-0">
           <div className="grid w-full grid-cols-3 gap-4">
@@ -224,7 +224,11 @@ const Dashboard = () => {
           </PageCard>
         </PageRow>
       </div>
-      <ContentOverview data={DATA_ACTIVE_PROGRAMME} columns={COLUMN_ACTIVE_PROGRAMME} />
+      <ContentOverview
+        dataTable={DATA_ACTIVE_PROGRAMME}
+        columns={COLUMN_ACTIVE_PROGRAMME}
+        titleTable={"ACTIVE COUNTRIES"}
+      />
     </div>
   );
 };
